boletim, topbar refactoring #2
