+++ client/src/pages/boletim/style.css
@@ -1,5 +1,6 @@
-.frequencia_title {
+.progresso_title {
     color: white;
+    margin-bottom: 2vh;
 }
 
 /* .boletimContainer {
