--- client/src/utils/routes/style.css
@@ -1,9 +1,0 @@
-.navbarContent {
- width: 100%;
- display: flex;
- justify-content: center
-}
-
-.components {
- margin-top: 7rem;
-}
